WIP, move game_internal to new classic ruleset
