+++ src/position/mod.rs
@@ -1,4 +1,2 @@
-mod position_api;
-mod position_internal;
 pub(crate) mod position_struct;
 pub(crate) mod util;
